# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Flow.",
        "title": "Flow API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/boards/{board_id}/invites": {
            "get": {
                "security": [
                    {
                        "jwt_auth": []
                    }
                ],
                "description": "Get invitation links to the board with ID with parameters (user must be author of the board)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board sharing [author]"
                ],
                "summary": "Get links",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the board",
                        "name": "board_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link list has been successfully fetched",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/definitions/domain.LinkParams"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - access denied",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "404": {
                        "description": "Board or links not found",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "jwt_auth": []
                    }
                ],
                "description": "Create invitation link to the board with parameters (person, time limit, usage limit) (user must be author of the board)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Board sharing [author]"
                ],
                "summary": "Create link",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the board",
                        "name": "board_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Link parameters: usernames for personal invitation; time limit for link activity; usage limit",
                        "name": "_",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/domain.Invitaion"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Link has been successfully created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "link": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "207": {
                        "description": "Link has been successfully created for valid names; Invalid usernames are returned",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/rest.ServerResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "invalid": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "link": {
                                                    "type": "string"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request parameters",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - access denied",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "404": {
                        "description": "Link not found",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/rest.ServerResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.Invitaion": {
            "type": "object",
            "properties": {
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time_limit": {
                    "type": "string"
                },
                "usage_limit": {
                    "type": "integer"
                }
            }
        },
        "domain.LinkParams": {
            "type": "object",
            "properties": {
                "link": {
                    "type": "string"
                },
                "names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "time_limit": {
                    "type": "string"
                },
                "usage_count": {
                    "type": "integer"
                },
                "usage_limit": {
                    "type": "integer"
                }
            }
        },
...
        "rest.ServerResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "description": {
                    "type": "string"
                }
            }
        }
    }
}